+++ __tests__/Builder/makeEnhancedNode.test.js
@@ -94,6 +94,7 @@ describe("make an EnhancedNode out of POJOs", () => {
     // const results: Result[] = await engine.mergeEnhancedNodes([enode]);
     // expect(results[0].getData()[0].isWritten()).toEqual(true);
   });
+
   test("deep enode", async () => {
     /* (:Person { NAME: "Rob" })-[:HAS_FRIEND]->(:Person { NAME: "Charlie" })-[:HAS_ROOMMATE]->(:Person { NAME: "Frank" }) */
     /* order is important */
@@ -113,6 +114,7 @@ describe("make an EnhancedNode out of POJOs", () => {
     expect(enode).toBeInstanceOf(EnhancedNode);
     expect(enode.getParticipatingRelationships()).toHaveLength(2);
   });
+  
   // test.only("deep enode", async () => {
   //   // (:Person { NAME: "Rob" })-[:HAS_FRIEND]->(:Person { NAME: "Charlie" })-[:HAS_ROOMMATE]->(:Person { NAME: "Frank" })
   //   const mn = builder.makeNode.bind(builder);